backend/requirements.txt
aniso8601==6.0.0
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2==2.9.6
python-decouple==3.8
python-dotenv==1.0.0
pytz==2019.1
six==1.12.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
Werkzeug==2.3.4
